nix installs: check if already installed

+++ 2-operating-system-and-linux/virtualization/arch-desktop/provisioner/nix-home-manager.sh
@@ -1,22 +1,29 @@
 #!/bin/bash
 set -e
 
-# Install nix home-manager per
-# https://nix-community.github.io/home-manager/index.xhtml
+# Check if home-manager already installed, if so skip 
+if command -v home-manager > /dev/null; then
 
-# Create a folder $HOME/.config/home-manager
-mkdir -p /home/vagrant/.config/home-manager
+    # Install nix home-manager per
+    # https://nix-community.github.io/home-manager/index.xhtml
 
-# Install home-manager
-nix-channel --add https://github.com/nix-community/home-manager/archive/master.tar.gz home-manager
-nix-channel --update
-nix-shell '<home-manager>' -A install
+    # Create a folder $HOME/.config/home-manager
+    mkdir -p /home/vagrant/.config/home-manager
 
-# Remove default and replace with custom home.nix
-rm /home/vagrant/.config/home-manager/home.nix
+    # Install home-manager
+    nix-channel --add https://github.com/nix-community/home-manager/archive/master.tar.gz home-manager
+    nix-channel --update
+    nix-shell '<home-manager>' -A install
 
-# Symlink /vagrant/home.nix to $HOME/.config/home-manager/home.nix
-ln -s /vagrant/home.nix /home/vagrant/.config/home-manager/home.nix
+    # Remove default and replace with custom home.nix
+    rm /home/vagrant/.config/home-manager/home.nix
+
+    # Symlink /vagrant/home.nix to $HOME/.config/home-manager/home.nix
+    ln -s /vagrant/home.nix /home/vagrant/.config/home-manager/home.nix
+
+fi
 
 # Install programs in home.nix and backup any existing files
 home-manager switch -b backup
+
+echo -e "-- home-manager Installed and Next generation created"
